extract segments list item into own module (and add new folder within right-border-tabs)
