+++ frontend/javascripts/oxalis/view/right-border-tabs/segments_tab/segments_view_helper.js
